pyproject.toml
[tool.poetry]
name = "mai2_pcount"
version = "0.1.0"
description = "wmc的机厅在线人数信息查询、修改插件~"
authors = ["Miaowing <shengwang52005@163.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/shengwang52005/mai2_pcount"
repository = "https://github.com/shengwang52005/mai2_pcount"

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = ">=2.3.2"
nonebot-adapter-onebot = ">=2.3.1"
httpx = ">=0.27.0"
pil_utils = "0.1.10"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
